+++ pkg/autoscaler/stats_scraper.go
@@ -18,7 +18,6 @@ package autoscaler
 
 import (
 	"fmt"
-	"io"
 	"net/http"
 	"time"
 
@@ -27,8 +26,6 @@ import (
 	"github.com/knative/serving/pkg/reconciler/v1alpha1/autoscaling/kpa/resources/names"
 	"github.com/knative/serving/pkg/resources"
 	"github.com/pkg/errors"
-	dto "github.com/prometheus/client_model/go"
-	"github.com/prometheus/common/expfmt"
 	corev1listers "k8s.io/client-go/listers/core/v1"
 )
 
@@ -49,6 +46,13 @@ type StatsScraper interface {
 	Scrape() (*StatMessage, error)
 }
 
+// scrapeClient defines the interface for collecting Revision metrics for a given
+// URL. Internal used only.
+type scrapeClient interface {
+	// Scrape scrapes the given URL.
+	Scrape(url string) (*Stat, error)
+}
+
 // cacheDisabledClient is a http client with cache disabled. It is shared by
 // every goruntime for a revision scraper.
 var cacheDisabledClient = &http.Client{
@@ -64,7 +68,7 @@ var cacheDisabledClient = &http.Client{
 // https://kubernetes.io/docs/concepts/services-networking/network-policies/
 // for details.
 type ServiceScraper struct {
-	httpClient          *http.Client
+	sClient             scrapeClient
 	endpointsLister     corev1listers.EndpointsLister
 	url                 string
 	namespace           string
@@ -75,21 +79,25 @@ type ServiceScraper struct {
 // NewServiceScraper creates a new StatsScraper for the Revision which
 // the given Metric is responsible for.
 func NewServiceScraper(metric *Metric, endpointsLister corev1listers.EndpointsLister) (*ServiceScraper, error) {
-	return newServiceScraperWithClient(metric, endpointsLister, cacheDisabledClient)
+	sClient, err := newHTTPScrapeClient(cacheDisabledClient)
+	if err != nil {
+		return nil, err
+	}
+	return newServiceScraperWithClient(metric, endpointsLister, sClient)
 }
 
 func newServiceScraperWithClient(
 	metric *Metric,
 	endpointsLister corev1listers.EndpointsLister,
-	httpClient *http.Client) (*ServiceScraper, error) {
+	sClient scrapeClient) (*ServiceScraper, error) {
 	if metric == nil {
 		return nil, errors.New("metric must not be nil")
 	}
 	if endpointsLister == nil {
 		return nil, errors.New("endpoints lister must not be nil")
 	}
-	if httpClient == nil {
-		return nil, errors.New("HTTP client must not be nil")
+	if sClient == nil {
+		return nil, errors.New("scrape client must not be nil")
 	}
 	revName := metric.Labels[serving.RevisionLabelKey]
 	if revName == "" {
@@ -98,7 +106,7 @@ func newServiceScraperWithClient(
 
 	serviceName := names.MetricsServiceName(revName)
 	return &ServiceScraper{
-		httpClient:          httpClient,
+		sClient:             sClient,
 		endpointsLister:     endpointsLister,
 		url:                 fmt.Sprintf("http://%s.%s:%d/metrics", serviceName, metric.Namespace, v1alpha1.RequestQueueMetricsPort),
 		metricKey:           NewMetricKey(metric.Namespace, metric.Name),
@@ -119,7 +127,7 @@ func (s *ServiceScraper) Scrape() (*StatMessage, error) {
 		return nil, nil
 	}
 
-	stat, err := s.scrapeViaURL()
+	stat, err := s.sClient.Scrape(s.url)
 	if err != nil {
 		return nil, err
 	}
@@ -146,71 +154,4 @@ func (s *ServiceScraper) Scrape() (*StatMessage, error) {
 		Stat: extrapolatedStat,
 		Key:  s.metricKey,
 	}, nil
-}
-
-func (s *ServiceScraper) scrapeViaURL() (*Stat, error) {
-	req, err := http.NewRequest(http.MethodGet, s.url, nil)
-	if err != nil {
-		return nil, err
-	}
-	resp, err := s.httpClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode < http.StatusOK || resp.StatusCode >= http.StatusMultipleChoices {
-		return nil, fmt.Errorf("GET request for URL %q returned HTTP status %v", s.url, resp.StatusCode)
-	}
-
-	return extractData(resp.Body)
-}
-
-func extractData(body io.Reader) (*Stat, error) {
-	var parser expfmt.TextParser
-	metricFamilies, err := parser.TextToMetricFamilies(body)
-	if err != nil {
-		return nil, fmt.Errorf("reading text format failed: %v", err)
-	}
-
-	now := time.Now()
-	stat := Stat{
-		Time: &now,
-	}
-
-	if pMetric := prometheusMetric(metricFamilies, "queue_average_concurrent_requests"); pMetric != nil {
-		stat.AverageConcurrentRequests = *pMetric.Gauge.Value
-	} else {
-		return nil, errors.New("could not find value for queue_average_concurrent_requests in response")
-	}
-
-	if pMetric := prometheusMetric(metricFamilies, "queue_average_proxied_concurrent_requests"); pMetric != nil {
-		stat.AverageProxiedConcurrentRequests = *pMetric.Gauge.Value
-	} else {
-		return nil, errors.New("could not find value for queue_average_proxied_concurrent_requests in response")
-	}
-
-	if pMetric := prometheusMetric(metricFamilies, "queue_operations_per_second"); pMetric != nil {
-		stat.RequestCount = int32(*pMetric.Gauge.Value)
-	} else {
-		return nil, errors.New("could not find value for queue_operations_per_second in response")
-	}
-
-	if pMetric := prometheusMetric(metricFamilies, "queue_proxied_operations_per_second"); pMetric != nil {
-		stat.ProxiedRequestCount = int32(*pMetric.Gauge.Value)
-	} else {
-		return nil, errors.New("could not find value for queue_proxied_operations_per_second in response")
-	}
-
-	return &stat, nil
-}
-
-// prometheusMetric returns the point of the first Metric of the MetricFamily
-// with the given key from the given map. If there is no such MetricFamily or it
-// has no Metrics, then returns nil.
-func prometheusMetric(metricFamilies map[string]*dto.MetricFamily, key string) *dto.Metric {
-	if metric, ok := metricFamilies[key]; ok && len(metric.Metric) > 0 {
-		return metric.Metric[0]
-	}
-
-	return nil
 }
